Use multilint
And fix all isort/modernize violations.
