+++ patchy/api.py
@@ -1,4 +1,6 @@
 # -*- encoding:utf-8 -*-
+from __future__ import absolute_import, division, print_function, unicode_literals
+
 import inspect
 import os
 import shutil
