+++ patchy/cache.py
@@ -1,5 +1,5 @@
-# -*- coding:utf-8 -*-
-from __future__ import division
+# -*- encoding:utf-8 -*-
+from __future__ import absolute_import, division, print_function, unicode_literals
 
 import random
