+++ requirements.txt
@@ -3,6 +3,8 @@ docutils==0.12
 flake8==2.6.0
 isort==4.2.5
 mccabe==0.5.0
+modernize==0.5
+multilint==1.0.0
 py==1.4.31
 pycodestyle==2.0.0
 pyflakes==1.2.3
