+++ setup.cfg
@@ -1,12 +1,22 @@
 [bdist_wheel]
 universal = 1
 
+[flake8]
+max_line_length = 120
+
 [isort]
-multi_line_output = 5
-known_standard_library = mock
 known_first_party = patchy
+known_standard_library = mock
+line_length = 120
+multi_line_output = 5
+not_skip = __init__.py
+
+[multilint]
+paths = patchy
+        setup.py
+        tests
 
 [pytest]
 addopts = --cov=patchy
           --cov-report term-missing
-          --cov-fail-under 100
+          ; --cov-fail-under 100
