+++ tests/base.py
@@ -1,3 +1,6 @@
+# -*- encoding:utf-8 -*-
+from __future__ import absolute_import, division, print_function, unicode_literals
+
 import unittest
 
 import patchy.api
