+++ tests/test_cache.py
@@ -1,5 +1,6 @@
-# -*- coding: utf-8 -*-
-from __future__ import division, print_function
+# -*- encoding:utf-8 -*-
+from __future__ import absolute_import, division, print_function, unicode_literals
+
 
 import pytest
