+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 envlist =
-    py35-codestyle,
+    py{27,35}-codestyle,
     py{27,34,35}
 
 [testenv]
@@ -8,8 +8,11 @@ setenv =
     PYTHONDONTWRITEBYTECODE=1
 install_command = pip install --no-deps {opts} {packages}
 deps = -r{toxinidir}/requirements.txt
-commands = ./runtests.py --nolint {posargs}
+commands = py.test {posargs}
 
+
+[testenv:py27-codestyle]
+commands = multilint
 
 [testenv:py35-codestyle]
-commands = ./runtests.py --lintonly
+commands = multilint
